feat: WolfApp (Main View) 생성
