+++ presentation/src/main/java/wolf/shin/earlyadopter/MainActivity.kt
@@ -6,7 +6,6 @@ import androidx.activity.compose.setContent
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
 import wolf.shin.earlyadopter.model.DockBarItem
 import wolf.shin.earlyadopter.navigation.BottomNavigationBar
 import wolf.shin.earlyadopter.navigation.InitNavigator
@@ -18,41 +17,8 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             AndroidEarlyAdopterTheme {
-
-                val navController = rememberNavController()
-
-                MainView(navController)
-
+                WolfApp()
             }
         }
-    }
-}
-
-@Composable
-private fun MainView(navController: NavHostController) {
-    Scaffold(
-        bottomBar = {
-            BottomNavigationBar(
-                items = listOf(
-                    DockBarItem(
-                        name = "Home",
-                        route = "Home",
-                        defaultIcon = R.drawable.ic_dockbar_off_home,
-                        selectedIcon = R.drawable.ic_dockbar_on_home
-                    ),
-                    DockBarItem(
-                        name = "News",
-                        route = "News",
-                        defaultIcon = R.drawable.ic_dockbar_off_news,
-                        selectedIcon = R.drawable.ic_dockbar_on_news
-                    ),
-                ),
-                navController = navController,
-                onItemClick = {
-                    navController.navigate(it.route)
-                })
-        }
-    ) {
-        InitNavigator(navController = navController)
     }
 }
